more columns on utxos, organize setup code
