+++ src/controllers/queries.ts
@@ -10,13 +10,6 @@ import { Op } from 'sequelize'
 import fetch from 'node-fetch'
 import * as helpers from '../helpers'
 
-/*
-  make the keysend go over the just-created channel
-  add columns: txid for the onchain payment
-    commit_fee, local_reserve, remote_reserve, hardcoded AMT
-*/
-
-
 type QueryType = 'onchain_address'
 export interface Query {
   type: QueryType
@@ -27,18 +20,18 @@ export interface Query {
 
 let queries: { [k: string]: Query } = {}
 
-let hub_pubkey = ''
+let hub_pubkey = '02290714deafd0cb33d2be3b634fc977a98a9c9fa1dd6c53cf17d99b350c08c67b'
 
-const hub_url = 'https://hub.sphinx.chat/api/v1/'
-async function get_hub_pubkey(){
-  const r = await fetch(hub_url+'/routingnode')
-  const j = await r.json()
-  if(j && j.pubkey) {
-    console.log("=> GOT HUB PUBKEY", j.pubkey)
-    hub_pubkey = j.pubkey
-  }
-}
-get_hub_pubkey()
+// const hub_url = 'https://hub.sphinx.chat/api/v1/'
+// async function get_hub_pubkey(){
+//   const r = await fetch(hub_url+'/routingnode')
+//   const j = await r.json()
+//   if(j && j.pubkey) {
+//     console.log("=> GOT HUB PUBKEY", j.pubkey)
+//     hub_pubkey = j.pubkey
+//   }
+// }
+// get_hub_pubkey()
 
 interface Accounting {
   id: number
@@ -49,6 +42,7 @@ interface Accounting {
   sourceApp: string
   date: string
   fundingTxid: string
+  onchainTxid: string
 }
 
 async function getReceivedAccountings(): Promise<Accounting[]> {
@@ -75,6 +69,7 @@ async function getPendingAccountings(): Promise<Accounting[]> {
   accountings.forEach(a => {
     const utxo = utxos.find(u => u.address === a.onchainAddress)
     if (utxo) {
+      const onchainTxid = utxo.outpoint && utxo.outpoint.txid_str
       ret.push(<Accounting>{
         id: a.id,
         pubkey: a.pubkey,
@@ -83,6 +78,7 @@ async function getPendingAccountings(): Promise<Accounting[]> {
         confirmations: utxo.confirmations,
         sourceApp: a.sourceApp,
         date: a.date,
+        onchainTxid: onchainTxid
       })
     }
   })
@@ -127,6 +123,7 @@ async function genChannelAndConfirmAccounting(acc: Accounting) {
     await models.Accounting.update({
       status: constants.statuses.received,
       fundingTxid: fundingTxid,
+      onchainTxid: acc.onchainTxid,
       amount: acc.amount
     }, {
       where: { id: acc.id }
@@ -177,8 +174,11 @@ async function checkChannelsAndKeysend(rec: Accounting){
       const msg: { [k: string]: any } = {
         type: constants.message_types.keysend,
       }
-      const MINUS_AMT = 2000
-      const amount = rec.amount - parseInt(chan.local_chan_reserve_sat||0) - parseInt(chan.remote_chan_reserve_sat||0) - parseInt(chan.commit_fee||0) - MINUS_AMT
+      const extraAmount = 2000
+      const localReserve = parseInt(chan.local_chan_reserve_sat||0)
+      const remoteReserve = parseInt(chan.remote_chan_reserve_sat||0)
+      const commitFee = parseInt(chan.commit_fee||0)
+      const amount = rec.amount - localReserve - remoteReserve - commitFee - extraAmount
       console.log('[WATCH] amt to final keysend', amount)
       helpers.performKeysendMessage({
         sender: owner,
@@ -188,7 +188,8 @@ async function checkChannelsAndKeysend(rec: Accounting){
           console.log('[WATCH] complete! Updating accounting, id:', rec.id)
           models.Accounting.update({
             status: constants.statuses.confirmed,
-            chanId: chan.chan_id
+            chanId: chan.chan_id,
+            extraAmount, localReserve, remoteReserve, commitFee
           }, {
             where: { id: rec.id }
           })
